bbs/requirements.txt
Django==2.0
djangorestframework==3.7.3
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
mysqlclient==1.3.12
pytz==2018.9
redis==3.2.0